Don't update an order in the redirect callback.
This creates a payment that doesn't need to be there. The responsiblity
of this function is to redirect to paypal for payment if necessary.
After processing the payment paypal redirects the user to CheckoutController#paypal_finish
which creates the appropriate payment.

The creation of a payment will invalidate all previous payments, which
was what was occuring for the payment created here.

Change-Id: I8d1675375a36a52717268c404a4fed832309441b